requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-docx
requests
python-multipart
